+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## [0.17.3] - 
 
+### Added
+- Support for RGB images in `asilib.Imager.keogram()`.
+
 ### Changed
 - Simplified `asilib.Imager.__getitem__()` to repay the technical debt. Before the method was difficult to reason about and it processed the [time] and [start_time:end_time] slice cases separately. Now `__getitem__` handles both of those cases in the same way.
+- Simplified how the `Imager._keogram` array is allocated when `asilib.Imager.keogram()` is called.
 
 ## [0.17.2] - 2023-06-28 
 ### Added